+++ MatrixSDK/MXRestClient.h
@@ -1184,35 +1184,6 @@ typedef MXHTTPOperation* (^MXRestClientIdentityServerAccessTokenHandler)(void (^
                        success:(void (^)(MXCreateRoomResponse *response))success
                        failure:(void (^)(NSError *error))failure NS_SWIFT_UNAVAILABLE("TEST");
 
-
-/**
- Create a room.
-
- @param name (optional) the room name.
- @param visibility (optional) the visibility of the room in the current HS's room directory.
- @param roomAlias (optional) the room alias on the home server the room will be created.
- @param topic (optional) the room topic.
- @param inviteArray (optional) A list of user IDs to invite to the room. This will tell the server to invite everyone in the list to the newly created room.
- @param invite3PIDArray (optional) A list of objects representing third party IDs to invite into the room.
- @param isDirect This flag makes the server set the is_direct flag on the m.room.member events sent to the users in invite and invite_3pid (Use NO by default).
- @param preset (optional) Convenience parameter for setting various default state events based on a preset.
-
- @param success A block object called when the operation succeeds. It provides a MXCreateRoomResponse object.
- @param failure A block object called when the operation fails.
-
- @return a MXHTTPOperation instance.
- */
-- (MXHTTPOperation*)createRoom:(NSString*)name
-                    visibility:(MXRoomDirectoryVisibility)visibility
-                     roomAlias:(NSString*)roomAlias
-                         topic:(NSString*)topic
-                        invite:(NSArray<NSString*>*)inviteArray
-                    invite3PID:(NSArray<MXInvite3PID*>*)invite3PIDArray
-                      isDirect:(BOOL)isDirect
-                        preset:(MXRoomPreset)preset
-                       success:(void (^)(MXCreateRoomResponse *response))success
-                       failure:(void (^)(NSError *error))failure NS_REFINED_FOR_SWIFT;
-
 /**
  Create a room.